move sandbox to dedicated package
